Mu.Rpc

Description

This module defines a type-level language to describe RPC-like microservices independently of the transport and protocol.

Documentation

type Package' = Package Symbol Symbol Symbol

Packages whose names are given by type-level strings.

data Package serviceName methodName argName

A package is a set of services.

Constructors

Package (Maybe serviceName) [Service serviceName methodName argName]

type Service' = Service Symbol Symbol Symbol

Services whose names are given by type-level strings.

data Service serviceName methodName argName

A service is a set of methods.

Constructors

Service serviceName [ServiceAnnotation] [Method serviceName methodName argName]

type Object = Service

An object is a set of fields, in GraphQL lingo.

type ServiceAnnotation = Type

Annotations for services. At this moment, such annotations can be of any type.

type Method' = Method Symbol Symbol Symbol

Methods whose names are given by type-level strings.

data Method serviceName methodName argName

A method is defined by its name, arguments, and return type.

Constructors

Method methodName [ServiceAnnotation] [Argument serviceName argName] (Return serviceName)

type ObjectField = Method

A field in an object takes some input objects, and returns a value or some other object, in GraphQL lingo.

type family LookupService (ss :: [Service snm mnm anm]) (s :: snm) :: Service snm mnm anm where ...

Equations

LookupService '[] s = TypeError (Text "could not find method " :<>: ShowType s)

LookupService (Service s anns ms ': ss) s = Service s anns ms

LookupService (other ': ss) s = LookupService ss s

type family LookupMethod (s :: [Method snm mnm anm]) (m :: mnm) :: Method snm mnm anm where ...

Look up a method in a service definition using its name.

Equations

LookupMethod '[] m = TypeError (Text "could not find method " :<>: ShowType m)

LookupMethod (Method m anns args r ': ms) m = Method m anns args r

LookupMethod (other ': ms) m = LookupMethod ms m

data TypeRef serviceName where

Constructors

PrimitiveRef :: Type -> TypeRef serviceName
A primitive type.

ObjectRef :: serviceName -> TypeRef serviceName
Chain with another service.

SchemaRef :: Schema typeName fieldName -> typeName -> TypeRef serviceName
Point to schema.

RegistryRef :: Registry -> Type -> Nat -> TypeRef serviceName
Registry subject, type to convert to, and preferred serialization version

THRef :: Type -> TypeRef serviceName
To be used only during TH generation!

ListRef :: TypeRef serviceName -> TypeRef serviceName
Represents a list of values.

OptionalRef :: TypeRef serviceName -> TypeRef serviceName
Represents a possibly-missing value.

type Argument' = Argument Symbol Symbol

Arguments whose names are given by type-level strings.

data Argument serviceName argName where

Defines the way in which arguments are handled.

Constructors

ArgSingle :: Maybe argName -> [ServiceAnnotation] -> TypeRef serviceName -> Argument serviceName argName
Use a single value.

ArgStream :: Maybe argName -> [ServiceAnnotation] -> TypeRef serviceName -> Argument serviceName argName
Consume a stream of values.

data Return serviceName where

Defines the different possibilities for returning information from a method.

Constructors

RetNothing :: Return serviceName
Fire and forget.

RetSingle :: TypeRef serviceName -> Return serviceName
Return a single value.

RetStream :: TypeRef serviceName -> Return serviceName
Return a stream of values.

RetThrows :: TypeRef serviceName -> TypeRef serviceName -> Return serviceName
Return a value or an error.