Mu.Rpc

Description

This module defines a type-level language to describe RPC-like microservices independently of the transport and protocol.

Documentation

type Package' = Package Symbol Symbol Symbol (TypeRef Symbol)

Packages whose names are given by type-level strings.

data Package serviceName methodName argName tyRef

A package is a set of services.

Constructors

Package (Maybe serviceName) [Service serviceName methodName argName tyRef]

type Service' = Service Symbol Symbol Symbol (TypeRef Symbol)

Services whose names are given by type-level strings.

data Service serviceName methodName argName tyRef

A service is a set of methods.

Constructors

Service serviceName [Method serviceName methodName argName tyRef]

type Object = 'Service

An object is a set of fields, in GraphQL lingo.

type Method' = Method Symbol Symbol Symbol (TypeRef Symbol)

Methods whose names are given by type-level strings.

data Method serviceName methodName argName tyRef

A method is defined by its name, arguments, and return type.

Constructors

Method methodName [Argument serviceName argName tyRef] (Return serviceName tyRef)

type ObjectField = 'Method

A field in an object takes some input objects, and returns a value or some other object, in GraphQL lingo.

type family LookupService (ss :: [Service snm mnm anm tr]) (s :: snm) :: Service snm mnm anm tr where ...

Look up a service in a package definition using its name.

Equations

LookupService '[] s = TypeError ('Text "could not find method " :<>: 'ShowType s)

LookupService ('Service s ms ': ss) s = 'Service s ms

LookupService (other ': ss) s = LookupService ss s

type family LookupMethod (s :: [Method snm mnm anm tr]) (m :: mnm) :: Method snm mnm anm tr where ...

Look up a method in a service definition using its name.

Equations

LookupMethod '[] m = TypeError ('Text "could not find method " :<>: 'ShowType m)

LookupMethod ('Method m args r ': ms) m = 'Method m args r

LookupMethod (other ': ms) m = LookupMethod ms m

data TypeRef serviceName where

Defines a reference to a type, either primitive or coming from the schema. TypeRefs are used to define arguments and result types.

Constructors

PrimitiveRef :: Type -> TypeRef serviceName
A primitive type.

ObjectRef :: serviceName -> TypeRef serviceName
Chain with another service.

SchemaRef :: Schema typeName fieldName -> typeName -> TypeRef serviceName
Point to schema.

RegistryRef :: Registry -> Type -> Nat -> TypeRef serviceName
Registry subject, type to convert to, and preferred serialization version

THRef :: Type -> TypeRef serviceName
To be used only during TH generation!

ListRef :: TypeRef serviceName -> TypeRef serviceName
Represents a list of values.

OptionalRef :: TypeRef serviceName -> TypeRef serviceName
Represents a possibly-missing value.

type Argument' = Argument Symbol Symbol (TypeRef Symbol)

Arguments whose names are given by type-level strings.

data Argument serviceName argName tyRef where

Defines the way in which arguments are handled.

Constructors

ArgSingle :: Maybe argName -> tyRef -> Argument serviceName argName tyRef
Use a single value.

ArgStream :: Maybe argName -> tyRef -> Argument serviceName argName tyRef
Consume a stream of values.

data Return serviceName tyRef where

Defines the different possibilities for returning information from a method.

Constructors

RetNothing :: Return serviceName tyRef
Fire and forget.

RetSingle :: tyRef -> Return serviceName tyRef
Return a single value.

RetStream :: tyRef -> Return serviceName tyRef
Return a stream of values.

RetThrows :: tyRef -> tyRef -> Return serviceName tyRef
Return a value or an error.

data TyInfo

Constructors

TyList TyInfo

TyOption TyInfo

TyTy Text

data RpcInfo i

Reflection

Constructors

NoRpcInfo

RpcInfo

Fields

packageInfo :: Package Text Text Text TyInfo

serviceInfo :: Service Text Text Text TyInfo

methodInfo :: Method Text Text Text TyInfo

headers :: RequestHeaders

extraInfo :: i

class ReflectRpcInfo (p :: Package') (s :: Service') (m :: Method') where

Methods

reflectRpcInfo :: Proxy p -> Proxy s -> Proxy m -> RequestHeaders -> i -> RpcInfo i